Move stuff from widgets/homework
