+++ app/lib/blackboard/show_delete_blackboard_item_dialog.dart
@@ -10,7 +10,7 @@ import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:sharezone/blackboard/blackboard_view.dart';
 import 'package:sharezone/main/application_bloc.dart';
-import 'package:sharezone/widgets/homework/delete_homework.dart';
+import 'package:sharezone/homework/shared/delete_homework.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
 
 import 'details/blackboard_details.dart';
