+++ app/lib/homework/student/src/homework_tile.dart
@@ -18,8 +18,8 @@ import 'package:sharezone/homework/homework_details/homework_details.dart';
 import 'package:sharezone/homework/homework_details/homework_details_view_factory.dart';
 import 'package:sharezone/submissions/homework_create_submission_page.dart';
 import 'package:sharezone/util/navigation_service.dart';
-import 'package:sharezone/widgets/homework/homework_card.dart';
-import 'package:sharezone/widgets/homework/homework_tile_template.dart';
+import 'package:sharezone/homework/shared/homework_card.dart';
+import 'package:sharezone/homework/shared/homework_tile_template.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
 
 enum HomeworkStatus { open, completed }
